requirements.txt
streamlit>=1.28.0
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.17.0
Pillow>=10.0.0

# Neo4j 数据库驱动
neo4j
